Word Reverese In Java
Write a Java program that uses a Vector to store a collection of words. The program should prompt the user to enter words until they enter the word "done" (case-insensitive). Afterward, display the words in reverse order, excluding the word "done." Ensure that the word "done" is not included in the final output.